Better solution for arrays
New function getIndices(): returns indices of identifier
f_name_ref()  returns the value of the array at the given index
Whenever assign is called, getIndices should be called on the identifier argument. If [] is returned, pass it in. That is the default args. That way you do not have to check whether or not you have indices